test: give the username-disclosure guard an always-built core entrypoint
The guard lived only in :ee-test, a conditionally-built module — an EE-less build (OSS fork, or the
CWD drift that drops :ee-app) shipped green with zero enforcement of what is a core invariant.

Extract the allowlist and the scan-and-assert into a single shared helper
(io.tolgee.testing.security.UsernameDisclosureGuard in :testing), keyed on string class names so
:testing needs no compile dependency on :api or spring-hateoas. Add a core entrypoint in :server-app
— built and tested in every build — anchored on the core allowlisted models, so core response
models stay guarded even when EE is absent. The existing EE test becomes a thin entrypoint that
additionally anchors an EE model. One allowlist, no duplication.
